MAINT isolate literals and move auth machinary to _account
