+++ microsetta_private_api/api/__init__.py
@@ -1,6 +1,6 @@
 from ._account import (
     find_accounts_for_login, register_account, claim_legacy_acct,
-    read_account, update_account, check_email_match,
+    read_account, update_account, check_email_match, verify_authrocket,
 )
 from ._consent import (
     render_consent_doc,
@@ -48,4 +48,5 @@ __all__ = [
     'read_kit',
     'associate_answered_survey',
     'submit_answered_survey',
+    'verify_authrocket',
 ]
