+++ microsetta_private_api/api/_account.py
@@ -1,9 +1,14 @@
 import uuid
 
+import jwt
 from flask import jsonify
+from jwt import InvalidTokenError
 
-from microsetta_private_api.api.implementation import JWT_ISS_CLAIM_KEY, \
-    JWT_SUB_CLAIM_KEY, JWT_EMAIL_CLAIM_KEY, _validate_account_access
+from werkzeug.exceptions import Unauthorized, Forbidden, NotFound
+
+from microsetta_private_api.api.literals import AUTHROCKET_PUB_KEY, \
+    INVALID_TOKEN_MSG, JWT_ISS_CLAIM_KEY, JWT_SUB_CLAIM_KEY, \
+    JWT_EMAIL_CLAIM_KEY, ACCT_NOT_FOUND_MSG
 from microsetta_private_api.model.account import Account, AuthorizationMatch
 from microsetta_private_api.model.address import Address
 from microsetta_private_api.repo.account_repo import AccountRepo
@@ -116,3 +121,61 @@ def update_account(account_id, body, token_info):
         t.commit()
 
         return jsonify(acc.to_api()), 200
+
+
+def verify_authrocket(token):
+    email_verification_key = 'email_verified'
+
+    try:
+        token_info = jwt.decode(token,
+                                AUTHROCKET_PUB_KEY,
+                                algorithms=["RS256"],
+                                verify=True,
+                                issuer="https://authrocket.com")
+    except InvalidTokenError as e:
+        raise(Unauthorized(INVALID_TOKEN_MSG, e))
+
+    if JWT_ISS_CLAIM_KEY not in token_info or \
+            JWT_SUB_CLAIM_KEY not in token_info or \
+            JWT_EMAIL_CLAIM_KEY not in token_info:
+        # token is malformed--no soup for you
+        raise Unauthorized(INVALID_TOKEN_MSG)
+
+    # if the user's email is not yet verified, they are forbidden to
+    # access their account even regardless of whether they have
+    # authenticated with authrocket
+    if email_verification_key not in token_info or \
+            token_info[email_verification_key] is not True:
+        raise Forbidden("Email is not verified")
+
+    return token_info
+
+
+def _validate_account_access(token_info, account_id):
+    with Transaction() as t:
+        account_repo = AccountRepo(t)
+        token_associated_account = account_repo.find_linked_account(
+            token_info['iss'],
+            token_info['sub'])
+        account = account_repo.get_account(account_id)
+        if account is None:
+            raise NotFound(ACCT_NOT_FOUND_MSG)
+        else:
+            # Whether or not the token_info is associated with an admin acct
+            token_authenticates_admin = \
+                token_associated_account is not None and \
+                token_associated_account.account_type == 'admin'
+
+            # Enum of how closely token info matches requested account_id
+            auth_match = account.account_matches_auth(
+                token_info[JWT_EMAIL_CLAIM_KEY],
+                token_info[JWT_ISS_CLAIM_KEY],
+                token_info[JWT_SUB_CLAIM_KEY])
+
+            # If token doesn't match requested account id, and doesn't grant
+            # admin access to the system, deny.
+            if auth_match == AuthorizationMatch.NO_MATCH and \
+                    not token_authenticates_admin:
+                raise Unauthorized()
+
+        return account
