+++ microsetta_private_api/api/_survey.py
@@ -1,7 +1,7 @@
 import flask
 from flask import jsonify
 
-from microsetta_private_api.api.implementation import \
+from microsetta_private_api.api._account import \
     _validate_account_access
 from microsetta_private_api.model.source import Source
 from microsetta_private_api.repo.source_repo import SourceRepo
